change UI in create&Edit Portfolio
