+++ frontend/src/App.tsx
@@ -4,12 +4,11 @@ import { Routes, Route } from 'react-router-dom';
 import Register from './pages/Register';
 import Home from './pages/Home';
 import Company from './pages/Company';
-import CreateCompanyProfile from './pages/CreateCompanyProfile';
 import Companys from './pages/Companys';
 import CreatePortfolio from './pages/CreatePortfolio';
 import Portfolio from './pages/Portfolio';
 import Login from './pages/Login';
-import ImageGallery from './components/Imgeselect';
+// import ImageGallery from './components/Imgeselect';
 import EditPortfolio from './pages/EditPortfolio';
 
 function App() {
@@ -17,7 +16,7 @@ function App() {
     <>
       <Toaster position="top-center" reverseOrder={false} />
       <Routes>
-        <Route path="/testimage" element={<ImageGallery />} />
+        {/* <Route path="/testimage" element={<ImageGallery />} /> */}
         {/* <Route path="/" element={<Homes3 />} /> */}
         <Route>
           <Route path="/" element={<Home />} />
@@ -25,7 +24,7 @@ function App() {
           <Route path="/register" element={<Register />} />
           <Route path="/companys" element={<Companys />} />
           <Route path="/company/:id" element={<Company />} />
-          <Route path="/company/create" element={<CreateCompanyProfile />} />
+          {/* <Route path="/company/create" element={<CreateCompanyProfile />} /> */}
           {/* <Route path="/createprofile" element={<CreateProfile />} /> */}
           <Route path="/portfolio/:portId" element={<Portfolio />} />
           <Route path="/portfolio/create" element={<CreatePortfolio />} />
