+++ frontend/src/components/Imgeselect.tsx
@@ -1,27 +1,27 @@
-// App.tsx
-import React from 'react';
-import ImageGallery from '../components/Showsileimg';
-import imag1 from '../img/phil-hearing-IYfp2Ixe9nM-unsplash.jpg';
-import imag2 from '../img/ravi-avaala-mR1CIDduGLc-unsplash.jpg';
-import imag3 from '../img/todd-kent-178j8tJrNlc-unsplash.jpg';
+// // App.tsx
+// import React from 'react';
+// import ImageGallery from '../components/Showsileimg';
+// import imag1 from '../img/phil-hearing-IYfp2Ixe9nM-unsplash.jpg';
+// import imag2 from '../img/ravi-avaala-mR1CIDduGLc-unsplash.jpg';
+// import imag3 from '../img/todd-kent-178j8tJrNlc-unsplash.jpg';
 
-const TestImage: React.FC = () => {
-  // Replace these image URLs with your own image URLs
-  const img =
-    'https://images.unsplash.com/photo-1554995207-c18c203602cb?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1170&q=80';
-  const images = [
-    img,
-    imag1,
-    imag2,
-    imag3,
-    // Add more image URLs here as needed
-  ];
+// const TestImage: React.FC = () => {
+//   // Replace these image URLs with your own image URLs
+//   const img =
+//     'https://images.unsplash.com/photo-1554995207-c18c203602cb?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1170&q=80';
+//   const images = [
+//     img,
+//     imag1,
+//     imag2,
+//     imag3,
+//     // Add more image URLs here as needed
+//   ];
 
-  return (
-    <div className="w-1/2">
-      <ImageGallery images={images} />
-    </div>
-  );
-};
+//   return (
+//     <div className="w-1/2">
+//       <ImageGallery images={images} />
+//     </div>
+//   );
+// };
 
-export default TestImage;
+// export default TestImage;
