+++ frontend/src/pages/CreateCompanyProfile.tsx
@@ -1,9 +1,9 @@
-import CreateCompanyProfile from '../components/CreateCompany';
+// import CreateCompanyProfile from '../components/CreateCompany';
 
-export default function CompanyProfile() {
-  return (
-    <>
-      <CreateCompanyProfile />
-    </>
-  );
-}
+// export default function CompanyProfile() {
+//   return (
+//     <>
+//       <CreateCompanyProfile />
+//     </>
+//   );
+// }
